HPE/requirements.txt
timm
yacs
ninja
progress
safetensors
torchmetrics